VMO-7102 rename resource editors
