+++ src/components/interaction-designer/resource-editors/index.ts
@@ -1,6 +1,7 @@
-export {default as HorizontalResourceEditor} from './HorizontalResourceEditor.vue'
-export {default as LanguageResourceEditor} from './LanguageResourceEditor.vue'
-export {default as ModeResourceEditor} from './ModeResourceEditor.vue'
+export {default as ResourceEditorCell} from './ResourceEditorCell.vue'
+export {default as PerLanguageResourceEditorRow} from './PerLanguageResourceEditorRow.vue'
+export {default as PerModeResourceEditorRow} from './PerModeResourceEditorRow.vue'
+export {default as PerLanguageResourceEditor} from './PerLanguageResourceEditor.vue'
+export {default as PerModeResourceEditor} from './PerModeResourceEditor.vue'
 export * from './ResourceEditor.model'
-export * from './ResourceEditor.vue'
 export * from './ResourceVariantTextEditor.vue'
